requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
pymongo
python-dotenv
